% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterFreq-data.R
\name{characterTable}
\alias{characterTable}
\title{These data are chinese character frequency Statistics based on read world news.
This is data to be included in my package.}
\format{\describe{
\item{character}{candidate}
\item{freq}{occurence of character}
\item{pct}{frequency of character}
}}
\usage{
data(characterTable)
}
\description{
These data are chinese character frequency Statistics based on read world news.
This is data to be included in my package.
}
\examples{
data(characterTable)
twoChar <- twoCharacters
wordFreq <- oneCharacters
}
\keyword{datasets}
