% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fiveChar}
\alias{fiveChar}
\title{Chinese Word Frequency with five Characters}
\format{A data frame with 3 variables:
\describe{
\item{\code{character}}{target character}
\item{\code{freq}}{occurence of target character}
\item{\code{pct}}{frequency of target character}
}}
\usage{
fiveChar
}
\description{
This dataframe inclues the usage frequecy and occurence of
Chinese word
}
\examples{
 fiveChar
}
\keyword{datasets}
