% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fourChar}
\alias{fourChar}
\title{Chinese Word Frequency with four Characters}
\format{A data frame with 3 variables:
\describe{
\item{\code{character}}{target character}
\item{\code{feq}}{occurence of target character}
\item{\code{pct}}{frequency of target character}
}}
\usage{
fourChar
}
\description{
This dataframe inclues the usage frequecy and occurence of
Chinese word
}
\examples{
 fourChar
}
\keyword{datasets}
