% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twoChar}
\alias{twoChar}
\title{Chinese Word Frequency with two Characters}
\format{A data frame with 3 variables:
\describe{
\item{\code{character}}{target character}
\item{\code{freq}}{occurence of target character}
\item{\code{pct}}{frequency of target character}
}}
\usage{
twoChar
}
\description{
This dataframe inclues the usage frequecy and occurence of
Chinese word
}
\examples{
 twoChar
}
\keyword{datasets}
