% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordMiner_previous.R
\name{wordMiner.previous}
\alias{wordMiner.previous}
\title{Occurrence and frequency of the previous word}
\usage{
wordMiner.previous(candidateWord, topN = 10)
}
\arguments{
\item{candidateWord}{A single Chinese character}

\item{topN}{The number of raws of output dataframe}
}
\value{
A dataframe containing the previous word and its occurrence and frequency
}
\description{
This is a function that lists the previous word appearing frequency
by a given chinese character based on real world
chinese character frequency Statistics table.
}
\examples{
wordMiner.previous(strsplit(levels(twoChar$character)[99], "")[[1]][1])
wordMiner.previous(strsplit(levels(twoChar$character)[99], "")[[1]][2], 15)

}
